feat[android]: move cloud download to own package
