+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/cloud/CloudDownloadMgr.java
@@ -1,4 +1,4 @@
-package com.tavultesoft.kmea.data;
+package com.tavultesoft.kmea.cloud;
 
 import android.app.DownloadManager;
 import android.content.BroadcastReceiver;
@@ -6,8 +6,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.net.Uri;
-
-import com.tavultesoft.kmea.cloud.CloudApiTypes;
 
 import java.io.File;
 import java.util.HashMap;
