+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/cloud/impl/CloudKeyboardDownloadReturns.java
@@ -1,4 +1,4 @@
-package com.tavultesoft.kmea.cloud;
+package com.tavultesoft.kmea.cloud.impl;
 
 import java.util.List;
 import java.util.Map;
